chapters/ch09_inference_frameworks/pyproject.toml
[project]
name = "ch09-inference-frameworks"
version = "0.1.0"
description = "Chapter 9: Production Inference Frameworks"
requires-python = ">=3.10"

dependencies = [
    "torch>=2.0",
    "numpy>=1.24",
    "transformers>=4.35",
    "pytest>=7.0",
    "requests>=2.28",
]

[project.optional-dependencies]
# Full dependencies for all labs
full = [
    "vllm>=0.2",
    "llama-cpp-python>=0.2",
    "sglang>=0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["lab01_huggingface/tests", "lab02_vllm/tests", "lab03_llamacpp/tests", "lab04_sglang/tests", "lab05_benchmark/tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v"
